Closes #57 Refactor all api and protection usage

+++ server/api/time-slots/[id].delete.ts
@@ -1,9 +1,16 @@
 import { serverSupabaseClient } from "#supabase/server";
 export default defineEventHandler(async (event) => {
   const businessId = await getBusinessId(event);
-  const client = await serverSupabaseClient<Database>(event);
 
   const id = getRouterParams(event).id;
+  if (!id) {
+    throw createError({
+      statusCode: 400,
+      statusMessage: "No se proporciono un ID",
+      message: "No se proporciono un ID de plantilla de horario",
+    });
+  }
+  const client = await serverSupabaseClient<Database>(event);
   const { error } = await client
     .from("time_slot_templates")
     .delete()
